# api/schema/openapi.yaml
openapi: 3.1.0
info:
  title: CMCEN / RCMCE API
  version: 1.0.0
  description: >
    OpenAPI-style inventory for the Express routes mounted by server/server.js.
    Schemas are intentionally broad where handlers return Mongoose documents.
    The existing developer actor may save blank descriptive profile/editorial
    fields. Ordinary actors retain required checks but may retain authorized
    blanks on later edits. Credentials, identifiers, references, valid nonblank
    values, dates and publication requirements remain enforced; see API ROUTES.md.
    Every /api endpoint is rate limited by source IP and may return 429 with
    Retry-After and RateLimit-* headers. Sensitive endpoints have stricter limits.
    Every response includes X-Request-ID for support correlation. Server failures
    are logged and audited without retaining request bodies or query strings.
servers:
  - url: http://localhost:3000
    description: Local development (Eric)
  - url: http://localhost:3001
    description: Local development (Bray)
  - url: https://cmcen-staging.corebot.ca
    description: Staging
  - url: https://beta.cmcen-rcmce.ca
    description: Beta
security:
  - bearerAuth: []
...
paths:
...
  /api/adopt-displays:
    get:
      operationId: listPublishedAdoptDisplays
      tags: [Pages]
      summary: List published bilingual display records, excluding staff IDs.
      security: []
      responses:
        '200':
          description: Object containing a displays array; drafts are excluded.
        '500':
          description: Catalogue unavailable.
  /api/admin/adopt-displays:
    get:
      operationId: listAdminAdoptDisplays
      tags: [Admin]
      summary: List all display records.
      x-permission: canManageAdoptDisplays
      responses:
        '200':
          description: Object containing a displays array.
        '401':
          description: Authentication required.
        '403':
          description: adopt_displays.manage required.
    post:
      operationId: createAdoptDisplay
      tags: [Admin]
      summary: Create a display; defaults to draft.
      x-permission: canManageAdoptDisplays
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdoptDisplayInput'
      responses:
        '201':
          description: Object containing the created display.
        '400':
          description: Invalid fields or missing title.
        '401':
          description: Authentication required.
        '403':
          description: adopt_displays.manage required.
  /api/admin/adopt-displays/{displayId}:
    parameters:
      - in: path
        name: displayId
        required: true
        schema:
          type: string
        description: MongoDB document ID, independent of display number.
    patch:
      operationId: updateAdoptDisplay
      tags: [Admin]
      summary: Update supplied fields; publish or unpublish explicitly.
      x-permission: canManageAdoptDisplays
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdoptDisplayInput'
      responses:
        '200':
          description: Object containing the saved display.
        '400':
          description: Invalid fields or missing title.
        '401':
          description: Authentication required.
        '403':
          description: adopt_displays.manage required.
        '404':
          description: Invalid ID or display not found.
    delete:
      operationId: deleteAdoptDisplay
      tags: [Admin]
      summary: Delete a display; preserve image assets.
      x-permission: canManageAdoptDisplays
      responses:
        '200':
          description: Display deleted.
        '401':
          description: Authentication required.
        '403':
          description: adopt_displays.manage required.
        '404':
          description: Invalid ID or display not found.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    AdoptDisplayInput:
      type: object
      description: At least one title language is required for ordinary saves and publication. A developer may save blank draft copy; existing authorized blanks may be retained. Bilingual fields replace both language values when supplied. Availability, amount and expiry are descriptive text.
      properties:
        displayNumber:
          type: string
          maxLength: 80
        imageUrl:
          type: string
          maxLength: 2000
          description: Empty or HTTP(S) URL without credentials.
        published:
          type: boolean
          default: false
        title:
          type: object
          properties:
            en:
              type: string
              maxLength: 240
            fr:
              type: string
              maxLength: 240
        description:
          type: object
          properties:
            en:
              type: string
              maxLength: 8000
            fr:
              type: string
              maxLength: 8000
        adoptionAmount:
          type: object
          properties:
            en:
              type: string
              maxLength: 240
            fr:
              type: string
              maxLength: 240
        availability:
          type: object
          properties:
            en:
              type: string
              maxLength: 240
            fr:
              type: string
              maxLength: 240
        recognition:
          type: object
          properties:
            en:
              type: string
              maxLength: 2000
            fr:
              type: string
              maxLength: 2000
        expiry:
          type: object
          properties:
            en:
              type: string
              maxLength: 240
            fr:
              type: string
              maxLength: 240
...